+++ src/StatsExp.cpp
@@ -182,7 +182,7 @@ Rcpp::NumericVector RcppSumNormalize(const Rcpp::NumericVector& vec,
 // [[Rcpp::export]]
 Rcpp::NumericVector RcppArithmeticSeq(double from, double to, int length_out) {
   // Call the CppArithmeticSeq function
-  std::vector<double> result = CppArithmeticSeq(from, to, length_out);
+  std::vector<double> result = CppArithmeticSeq(from, to, static_cast<size_t>(length_out));
   // Convert the result back to Rcpp::NumericVector
   return Rcpp::wrap(result);
 }
